requirements.txt
fastapi==0.109.1
uvicorn==0.15.0
pytest==6.2.5
isNullOrEmpty==1.0.0
boto3==1.28.0
PyJWT==2.8.0
python-multipart